Client: separated configs definitions into a separate file
